Implementalçao de filtro em multiplas obs ação e modificação do nome da pasta
