+++ Front/src/app/ponto-presenca/ponto-presenca.component.html
@@ -123,9 +123,9 @@
             <div class="two fields">
               <div class="field" [ngClass]="{'error': tipo_solicitacao.invalid && tipo_solicitacao.dirty}">
                 <label>Selecionar Solicitação</label>
-                <select class="ui fluid search dropdown" name="tipo_solicitacao" #tipo_solicitacao="ngModel" [(ngModel)]="mSolicitacoes.tipo_solicitacao" (change)="ActiveInputAnalise(mSolicitacoes.tipo_solicitacao)" required>
+                <select class="ui fluid search dropdown" name="tipo_solicitacao" #tipo_solicitacao="ngModel" [(ngModel)]="mSolicitacoes.tipo_solicitacao"  (change)="ActiveInputAnalise(mSolicitacoes.tipo_solicitacao)" required>
                   <option value="">Selecione</option>
-                  <option *ngFor="let multiplaSolicitacao of multiplasSolicitacoes" value="{{multiplaSolicitacao.tipo_solicitacao}}">{{multiplaSolicitacao.descricao}}</option>
+                  <option *ngFor="let multiplaSolicitacao of multiplasSolicitacoes" [ngValue]="multiplaSolicitacao">{{multiplaSolicitacao.descricao}}</option>
                 </select>
                 <div class="invalid-feedback" [ngClass]="{'active': tipo_solicitacao.invalid && tipo_solicitacao.dirty}">
                   Solicitação é obrigatória.
