+++ Front/src/app/ponto-presenca/ponto-presenca.modulo.ts
@@ -13,7 +13,7 @@ import { PontoPresencaComponent } from './ponto-presenca.component';
 import { PontoPresencaService } from './ponto-presenca.service';
 import { ContatoModule } from '../contato/contato.module';
 import { RouterModule } from '@angular/router';
-import { ObsAcaoComponent } from './ponto-presenca-obs-acao/ponto-presenca-obs-acao.component';
+import { ObsAcaoComponent } from './ponto-presenca-multiplas-obs-acao/ponto-presenca-multiplas-obs-acao.component';
 
 @NgModule({
     imports: [
